+++ PgRoutiner/Builder/Dump/PgDumpCache.cs
@@ -75,7 +75,8 @@ public static class PgDumpCache
         }
         if (Current.Value.DumpPgCommands)
         {
-            Program.WriteLine(ConsoleColor.White, $"{pgDumpCmd} {process.StartInfo.Arguments}");
+            if (Current.Value.Verbose)
+                Program.WriteLine(ConsoleColor.White, $"{pgDumpCmd} {process.StartInfo.Arguments}");
         }
         process.StartInfo.UseShellExecute = false;
         process.StartInfo.CreateNoWindow = true;